+++ tests/test_state_opt/test_ppt_distinguishability.py
@@ -37,8 +37,26 @@ def test_ppt_distinguishability_yyd_density_matrices():
     states = [rho_1, rho_2, rho_3, rho_4]
     probs = [1 / 4, 1 / 4, 1 / 4, 1 / 4]
 
-    res = ppt_distinguishability(states, probs)
-    np.testing.assert_equal(np.isclose(res, 7 / 8, atol=0.001), True)
+    # Min-error tests:
+    primal_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 7 / 8, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 7 / 8, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=probs, dist_method="unambiguous", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=probs, dist_method="unambiguous", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 3 / 4, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 3 / 4, atol=0.001), True)
 
 
 def test_ppt_distinguishability_yyd_vectors():
@@ -67,8 +85,25 @@ def test_ppt_distinguishability_yyd_vectors():
     states = [x_1, x_2, x_3, x_4]
     probs = [1 / 4, 1 / 4, 1 / 4, 1 / 4]
 
-    res = ppt_distinguishability(states, probs)
-    np.testing.assert_equal(np.isclose(res, 7 / 8, atol=0.001), True)
+    primal_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 7 / 8, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 7 / 8, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=probs, dist_method="unambiguous", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=probs, dist_method="unambiguous", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 3 / 4, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 3 / 4, atol=0.001), True)
 
 
 def test_ppt_distinguishability_yyd_states_no_probs():
@@ -102,8 +137,25 @@ def test_ppt_distinguishability_yyd_states_no_probs():
 
     states = [rho_1, rho_2, rho_3, rho_4]
 
-    res = ppt_distinguishability(states)
-    np.testing.assert_equal(np.isclose(res, 7 / 8, atol=0.001), True)
+    primal_res = ppt_distinguishability(
+        states, probs=None, dist_method="min-error", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=None, dist_method="min-error", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 7 / 8, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 7 / 8, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=None, dist_method="unambiguous", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=None, dist_method="unambiguous", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 3 / 4, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 3 / 4, atol=0.001), True)
 
 
 def test_ppt_distinguishability_werner_hiding_pairs():
@@ -149,8 +201,26 @@ def test_ppt_distinguishability_werner_hiding_pairs():
     states = [sigma_0, sigma_1]
 
     expected_val = 1 / 2 + 1 / (dim + 1)
-    res = ppt_distinguishability(states)
-    np.testing.assert_equal(np.isclose(res, expected_val, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=None, dist_method="min-error", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=None, dist_method="min-error", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, expected_val, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, expected_val, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=None, dist_method="unambiguous", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=None, dist_method="unambiguous", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, 1 / 3, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, 1 / 3, atol=0.001), True)
 
 
 def test_ppt_distinguishability_four_bell_states():
@@ -164,6 +234,9 @@ def test_ppt_distinguishability_four_bell_states():
     The resource state is defined by
 
     ..math::
+        |\tau_{\epsilon} \rangle = \sqrt{\frac{1+\epsilon}{2}} +
+        |0\rangle | 0\rangle +
+        \sqrt{\frac{1-\epsilon}{2}} |1 \rangle |1 \rangle
 
     The closed form probability with which Alice and Bob can distinguish via
     PPT measurements is given as follows
@@ -201,9 +274,17 @@ def test_ppt_distinguishability_four_bell_states():
     ]
     probs = [1 / 4, 1 / 4, 1 / 4, 1 / 4]
 
-    res = ppt_distinguishability(states, probs)
     exp_res = 1 / 2 * (1 + np.sqrt(1 - eps ** 2))
-    np.testing.assert_equal(np.isclose(res, exp_res, atol=0.001), True)
+
+    primal_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=True
+    )
+    dual_res = ppt_distinguishability(
+        states, probs=probs, dist_method="min-error", strategy=False
+    )
+
+    np.testing.assert_equal(np.isclose(primal_res, exp_res, atol=0.001), True)
+    np.testing.assert_equal(np.isclose(dual_res, exp_res, atol=0.001), True)
 
 
 if __name__ == "__main__":