+++ toqito/state_opt/ppt_distinguishability.py
@@ -9,31 +9,26 @@ from .state_helper import __is_states_valid, __is_probs_valid
 
 
 def ppt_distinguishability(
-    states: List[np.ndarray], probs: List[float] = None
+    states: List[np.ndarray],
+    probs: List[float] = None,
+    dist_method="min-error",
+    strategy=False,
 ) -> float:
     r"""
-    Compute probability of distinguishing a state via PPT measurements [COS13]_.
+    Compute probability of optimally distinguishing a state via PPT measurements [COS13]_.
 
     Implements the semidefinite program (SDP) whose optimal value is equal to the maximum
     probability of perfectly distinguishing orthogonal maximally entangled states using any PPT
     measurement; a measurement whose operators are positive under partial transpose. This SDP was
     explicitly provided in [COS13]_.
 
-    Specifically, the function implements the dual problem (as this is computationally more
-    efficient) and is defined as:
+    One can specify the distinguishability method using the :code:`dist_method` argument.
 
-    .. math::
-        \begin{equation}
-            \begin{aligned}
-                \text{minimize:} \quad & \frac{1}{k} \text{Tr}(Y) \\
-                \text{subject to:} \quad & Y - \rho_j \geq \text{T}_{\mathcal{A}} (Q_j),
-                                           \quad j = 1, \ldots, k, \\
-                                         & Y \in \text{Herm}(\mathcal{A} \otimes
-                                          \mathcal{B}), \\
-                                        & Q_1, \ldots, Q_k \in
-                                          \text{Pos}(\mathcal{A} \otimes \mathcal{B}).
-            \end{aligned}
-        \end{equation}
+    For :code:`dist_method = "min-error"`, this is the default method that yields the probability of
+    distinguishing quantum states via PPT measurements that minimize the probability of error.
+
+    For :code:`dist_method = "unambiguous"`, Alice and Bob never provide an incorrect answer,
+    although it is possible that their answer is inconclusive.
 
     Examples
     ==========
@@ -104,18 +99,19 @@ def ppt_distinguishability(
         Physical review letters 109.2 (2012): 020506.
         https://arxiv.org/abs/1107.3224
 
+    :param states: A list of states provided as either matrices or vectors.
+    :param probs: Respective list of probabilities each state is selected.
+    :param dist_method: Method of distinguishing to use.
+    :param strategy: Returns strategy if :code:`True` and does not otherwise.
     :return: The optimal probability with which the states can be distinguished
              via PPT measurements.
     """
-    constraints = []
-    meas = []
-
     __is_states_valid(states)
     if probs is None:
         probs = [1 / len(states)] * len(states)
     __is_probs_valid(probs)
 
-    dim_x, dim_y = states[0].shape
+    _, dim_y = states[0].shape
 
     # The variable `states` is provided as a list of vectors. Transform them
     # into density matrices.
@@ -123,21 +119,172 @@ def ppt_distinguishability(
         for i, state_ket in enumerate(states):
             states[i] = state_ket * state_ket.conj().T
 
-    y_var = cvxpy.Variable((dim_x, dim_x), hermitian=True)
-    objective = cvxpy.Minimize(cvxpy.trace(cvxpy.real(y_var)))
+    if strategy:
+        return primal_problem(states, probs, dist_method)
+    return dual_problem(states, probs, dist_method)
+
+
+def primal_problem(
+    states: List[np.ndarray], probs: List[float] = None, dist_method="min-error"
+) -> float:
+    r"""
+    Calculate primal problem for PPT distinguishability.
+
+    The minimum-error semidefinite program implemented is defined as:
+
+    .. math::
+    \begin{equation}
+        \begin{aligned}
+            \text{maximize:} \quad & \sum_{j=1}^k \langle P_j, \rho_j \rangle \\
+            \text{subject to:} \quad & P_1 + \cdots + P_k = \mathbb{I}_{\mathcal{A}}
+                                        \otimes \mathbb{I}_{\mathcal{B}}, \\
+                                     & P_1, \ldots, P_k \in \text{PPT}(\mathcal{A} : \mathcal{B}).
+        \end{aligned}
+    \end{equation}
+
+    The unambiguous semidefinite program implemented is defined as:
+
+    .. math::
+    \begin{equation}
+        \begin{aligned}
+            \text{maximize:} \quad & \sum_{j=1}^k \langle P_j, \rho_j \rangle \\
+            \text{subject to:} \quad & P_1 + \cdots + P_k = \mathbb{I}_{\mathcal{A}}
+                                        \otimes \mathbb{I}_{\mathcal{B}}, \\
+                                     & P_1, \ldots, P_k
+                                      \in \text{PPT}(\mathcal{A} : \mathcal{B}), \\
+                                     & \langle P_i, \rho_j \rangle = 0,
+                                       \quad 1 \leq i, j \leq k, \quad i \not= j.
+        \end{aligned}
+    \end{equation}
+
+    :param states: A list of states provided as either matrices or vectors.
+    :param probs: Respective list of probabilities each state is selected.
+    :param dist_method: Method of distinguishing to use.
+    :return: The optimal value of the PPT primal problem SDP.
+    """
+    dim_x, _ = states[0].shape
+
+    obj_func = []
+    meas = []
+    constraints = []
 
     dim = int(np.log2(dim_x))
     dim_list = [2] * int(np.log2(dim_x))
 
     sys_list = list(range(1, dim, 2))
-    if not sys_list:
-        sys_list = [2]
+
+    # Unambiguous consists of k + 1 operators, where the outcome of the k+1^st corresponds to the
+    # inconclusive answer.
+    if dist_method == "unambiguous":
+        for i in range(len(states) + 1):
+            meas.append(cvxpy.Variable((dim_x, dim_x), PSD=True))
+            constraints.append(partial_transpose(meas[i], sys_list, dim_list) >> 0)
+
+        for i, _ in enumerate(states):
+            for j, _ in enumerate(states):
+                if i != j:
+                    constraints.append(
+                        probs[j] * cvxpy.trace(states[j].conj().T @ meas[i]) == 0
+                    )
+
+    # Minimize error of distinguishing via PPT measurements.
+    elif dist_method == "min-error":
+        for i, _ in enumerate(states):
+            meas.append(cvxpy.Variable((dim_x, dim_x), PSD=True))
+            constraints.append(partial_transpose(meas[i], sys_list, dim_list) >> 0)
 
     for i, _ in enumerate(states):
+        obj_func.append(probs[i] * cvxpy.trace(states[i].conj().T @ meas[i]))
+
+    constraints.append(sum(meas) == np.identity(dim_x))
+
+    objective = cvxpy.Maximize(sum(obj_func))
+    problem = cvxpy.Problem(objective, constraints)
+    sol_default = problem.solve()
+
+    return sol_default
+
+
+def dual_problem(
+    states: List[np.ndarray], probs: List[float] = None, dist_method="min-error"
+) -> float:
+    r"""
+    Calculate dual problem for PPT distinguishability.
+
+    The minimum-error semidefinite program implemented is defined as:
+
+    .. math::
+    \begin{equation}
+        \begin{aligned}
+            \text{minimize:} \quad & \frac{1}{k} \text{Tr}(Y) \\
+            \text{subject to:} \quad & Y - \rho_j \geq \text{T}_{\mathcal{A}} (Q_j),
+                                        \quad j = 1, \ldots, k, \\
+                                     & Y \in \text{Herm}(\mathcal{A} \otimes
+                                        \mathcal{B}), \\
+                                     & Q_1, \ldots, Q_k \in
+                                        \text{Pos}(\mathcal{A} \otimes \mathcal{B}).
+        \end{aligned}
+    \end{equation}
+
+    The unambiguous semidefinite program implemented is defined as:
+
+    .. math::
+    \begin{equation}
+        \begin{aligned}
+            \text{minimize:} \quad & \frac{1}{k} \text{Tr}(Y) \\
+            \text{subject to:} \quad & Y - \rho_j + \sum_{\substack{i \leq i \leq k \\ i \not= j}}
+                                        y_{i,j} \rho_i \geq T_{\mathcal{A}}(Q_j),
+                                        \quad j = 1, \ldots, k, \\
+                                     & Y \geq T_{\mathcal{A}}(Q_{k+1}), \\
+                                     & Y \in \text{Herm}(\mathcal{A} \otimes
+                                        \mathcal{B}), \\
+                                     & Q_1, \ldots, Q_k \in
+                                        \text{Pos}(\mathcal{A} \otimes \mathcal{B}), \\
+                                     & y_{i,j} \in \mathcal{R}. \quad 1 \leq i, j \leq k,
+                                        \quad i \not= j.
+        \end{aligned}
+    \end{equation}
+
+    :param states: A list of states provided as either matrices or vectors.
+    :param probs: Respective list of probabilities each state is selected.
+    :param dist_method: Method of distinguishing to use.
+    :return: The optimal value of the PPT dual problem SDP.
+    """
+    constraints = []
+    meas = []
+
+    dim_x, _ = states[0].shape
+
+    y_var = cvxpy.Variable((dim_x, dim_x), hermitian=True)
+    objective = cvxpy.Minimize(cvxpy.trace(cvxpy.real(y_var)))
+
+    dim = int(np.log2(dim_x))
+    dim_list = [2] * int(np.log2(dim_x))
+    sys_list = list(range(1, dim, 2))
+
+    if dist_method == "min-error":
+        for i, _ in enumerate(states):
+            meas.append(cvxpy.Variable((dim_x, dim_x), PSD=True))
+            constraints.append(
+                cvxpy.real(y_var - probs[i] * states[i])
+                >> partial_transpose(meas[i], sys=sys_list, dim=dim_list)
+            )
+
+    if dist_method == "unambiguous":
+        for j, _ in enumerate(states):
+            sum_val = 0
+            for i, _ in enumerate(states):
+                if i != j:
+                    sum_val += cvxpy.real(cvxpy.Variable()) * probs[i] * states[i]
+            meas.append(cvxpy.Variable((dim_x, dim_x), PSD=True))
+            constraints.append(
+                cvxpy.real(y_var - probs[j] * states[j] + sum_val)
+                >> partial_transpose(meas[j], sys=sys_list, dim=dim_list)
+            )
+
         meas.append(cvxpy.Variable((dim_x, dim_x), PSD=True))
         constraints.append(
-            cvxpy.real(y_var - probs[i] * states[i])
-            >> partial_transpose(meas[i], sys=sys_list, dim=dim_list)
+            cvxpy.real(y_var) >> partial_transpose(meas[-1], sys=sys_list, dim=dim_list)
         )
 
     problem = cvxpy.Problem(objective, constraints)